requirements.txt
# Core dependencies
rich>=13.0.0
ddgs>=5.0.0
trafilatura>=1.7.0
google-genai>=0.2.0
google-auth>=2.36.0

# Existing dependencies (if any)
tavily-python>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
